# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Vigilância Participativa API
  description: API REST para sistema de vigilância participativa em saúde pública
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: https://api.example.com/v1
    description: Production server
  - url: https://staging-api.example.com/v1
    description: Staging server
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT no formato Bearer
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 100
        totalPages:
          type: integer
        totalItems:
          type: integer
    PaginationLinks:
      type: object
      properties:
        first:
          type: string
          format: uri
        last:
          type: string
          format: uri
        prev:
          type: string
          format: uri
          nullable: true
        next:
          type: string
          format: uri
          nullable: true
...
    Context:
      type: object
      required:
        - name
        - locationId
        - accessType
      properties:
        id:
          type: integer
          readOnly: true
        locationId:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        type:
          type: string
          nullable: true
        accessType:
          type: string
          enum: [PUBLIC, PRIVATE]
        active:
          type: boolean
          default: true
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
    ContextCreate:
      type: object
      required:
        - name
        - locationId
        - accessType
      properties:
        locationId:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        type:
          type: string
          nullable: true
        accessType:
          type: string
          enum: [PUBLIC, PRIVATE]
        active:
          type: boolean
          default: true
    ContextUpdate:
      type: object
      properties:
        locationId:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        type:
          type: string
          nullable: true
        accessType:
          type: string
          enum: [PUBLIC, PRIVATE]
        active:
          type: boolean
...
    ContextManager:
      type: object
      required:
        - userId
        - contextId
      properties:
        id:
          type: integer
          readOnly: true
        userId:
          type: integer
        contextId:
          type: integer
        active:
          type: boolean
          default: true
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
    ContextManagerCreate:
      type: object
      required:
        - userId
        - contextId
      properties:
        userId:
          type: integer
        contextId:
          type: integer
        active:
          type: boolean
          default: true
    ContextManagerUpdate:
      type: object
      properties:
        userId:
          type: integer
        contextId:
          type: integer
        active:
          type: boolean
...
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: ID do recurso
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Número da página
    PageSizeParam:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Tamanho da página
  responses:
    BadRequestError:
      description: Requisição inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnauthorizedError:
      description: Não autorizado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFoundError:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /contexts:
    get:
      tags:
        - Contexts
      summary: Listar contextos
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: active
          in: query
          schema:
            type: boolean
        - name: locationId
          in: query
          schema:
            type: integer
        - name: accessType
          in: query
          schema:
            type: string
            enum: [PUBLIC, PRIVATE]
      responses:
        '200':
          description: Lista de contextos
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Context'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
                  links:
                    $ref: '#/components/schemas/PaginationLinks'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags:
        - Contexts
      summary: Criar contexto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContextCreate'
      responses:
        '201':
          description: Contexto criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Context'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /contexts/{id}:
    get:
      tags:
        - Contexts
      summary: Obter contexto
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Detalhes do contexto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Context'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    put:
      tags:
        - Contexts
      summary: Atualizar contexto
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContextUpdate'
      responses:
        '200':
          description: Contexto atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Context'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    delete:
      tags:
        - Contexts
      summary: Deletar contexto
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Contexto deletado
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /contexts/{contextId}/managers:
    get:
      tags:
        - Context Managers
      summary: Listar managers do contexto
      parameters:
        - name: contextId
          in: path
          required: true
          schema:
            type: integer
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Lista de managers
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ContextManager'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
                  links:
                    $ref: '#/components/schemas/PaginationLinks'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags:
        - Context Managers
      summary: Adicionar manager ao contexto
      parameters:
        - name: contextId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContextManagerCreate'
      responses:
        '201':
          description: Manager adicionado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContextManager'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /contexts/{contextId}/managers/{id}:
    get:
      tags:
        - Context Managers
      summary: Obter manager
      parameters:
        - name: contextId
          in: path
          required: true
          schema:
            type: integer
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Detalhes do manager
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContextManager'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    put:
      tags:
        - Context Managers
      summary: Atualizar manager
      parameters:
        - name: contextId
          in: path
          required: true
          schema:
            type: integer
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContextManagerUpdate'
      responses:
        '200':
          description: Manager atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContextManager'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    delete:
      tags:
        - Context Managers
      summary: Remover manager do contexto
      parameters:
        - name: contextId
          in: path
          required: true
          schema:
            type: integer
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Manager removido
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...